requirements.txt
requests
pandas
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.10.0
pyarrow>=10.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0